Backend2/requirements.txt
Flask
pymongo
python-dotenv
google-genai  # for Gemini (stubbed for now)
google-generativeai